add serde by serialize/deserialize
