--- libs/ext2/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "ext2"
-version = "0.0.1"
-edition = "2021"
-
-[lib]
-name = "ext2"
-path = "lib.rs"
-
-[features]
-no_std = []
-
-[dependencies]
