+++ libs/ext2_fs/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "ext2_fs"
+version = "0.0.1"
+edition = "2021"
+
+[lib]
+name = "ext2_fs"
+path = "lib.rs"
+
+[features]
+no_std = []
+
+[dependencies]
+serde = { version = "1.0", default-features = false, features = ["derive"] }
+postcard = "0.7.2"
+serde-big-array = "0.3.2"
+heapless = "0.7.8"
