+++ libs/ext2_fs/error.rs
@@ -1,9 +1,7 @@
-use alloc::string::String;
-
 pub type Result<T> = core::result::Result<T, FileSysError>;
 
 /// Error definition
-#[derive(Clone, Debug, PartialEq )]
+#[derive(Clone, Debug, PartialEq)]
 pub enum FileSysError {
     EOF,
     ENOMEM
