+++ libs/ext2_fs/tools.rs
@@ -1,0 +1,2 @@
+pub type U32le = u32;
+pub type U16le = u16;
